requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
polars==0.20.21
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.4
